pyproject.toml
[tool.poetry]
name = "mrtools"
version = "0.2.0"
description = "Modern ROOT Tools for CMS Analysis using RDataFrames"
authors = ["Dietrich Liko <Dietrich.Liko@oeaw.ac.at>"]
readme = "README.md"
packages = [{include = "mrtools", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
ruamel-yaml = "^0.17.31"
correctionlib = "^2.2.2"
dask = "^2023.5.1"
distributed = "^2023.5.1"
dask_jobqueue = "^0.8.1"
# workaround
urllib3 = "<2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.5.9"
mypy = "^1.3.0"
reorder-python-imports = "^3.9.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.1"
sphinxcontrib-napoleon = "^0.7"

[tool.poetry.scripts]
mrtools = "mrtools.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
